// backend_node/src/bops.gen.ts
export enum RideableType {
	Unknown = 0,

	Bicycle = 1,

	Ebicycle = 2,
}

export enum MemberCasual {
	Unknown = 0,

	Member = 1,

	Casual = 2,
}

export interface ITrip extends BebopRecord {
	rideId?: string;

	rideableType?: RideableType;

	startedAt?: Date;

	endedAt?: Date;

	startStationName?: string;

	startStationId?: string;

	endStationName?: string;

	endStationId?: string;

	startLat?: number;

	startLng?: number;

	endLat?: number;

	endLng?: number;

	memberCasual?: MemberCasual;
}
// ...
export interface IServerResponseAll extends BebopRecord {
	trips?: Array<ITrip>;
}

export class ServerResponseAll implements IServerResponseAll {
	public trips?: Array<ITrip>;

	constructor(record: IServerResponseAll) {
		this.trips = record.trips;
	}

	/**
	 * Serializes the current instance into a JSON-Over-Bebop string
	 */
	public stringify(): string {
		return ServerResponseAll.encodeToJSON(this);
	}

	/**
	 * Serializes the specified object into a JSON-Over-Bebop string
	 */
	public static encodeToJSON(record: IServerResponseAll): string {
		return JSON.stringify(record, BebopJson.replacer);
	}

	/**
	 * Validates that the runtime types of members in the current instance are correct.
	 */
	public validateTypes(): void {
		ServerResponseAll.validateCompatibility(this);
	}

	/**
	 * Validates that the specified dynamic object can become an instance of {@link ServerResponseAll}.
	 */
	public static validateCompatibility(record: IServerResponseAll): void {
		if (record.trips !== undefined) {
			BebopTypeGuard.ensureArray(record.trips, Trip.validateCompatibility);
		}
	}

	/**
	 * Unsafely creates an instance of {@link ServerResponseAll} from the specified dynamic object. No type checking is performed.
	 */
	public static unsafeCast(record: any): IServerResponseAll {
		return new ServerResponseAll(record);
	}

	/**
	 * Creates a new {@link ServerResponseAll} instance from a JSON-Over-Bebop string. Type checking is performed.
	 */
	public static fromJSON(json: string): IServerResponseAll {
		if (typeof json !== "string" || json.trim().length === 0) {
			throw new BebopRuntimeError(`ServerResponseAll.fromJSON: expected string`);
		}
		const parsed = JSON.parse(json, BebopJson.reviver);
		ServerResponseAll.validateCompatibility(parsed);
		return ServerResponseAll.unsafeCast(parsed);
	}
	public encode(): Uint8Array {
		return ServerResponseAll.encode(this);
	}

	public static encode(record: IServerResponseAll): Uint8Array {
		const view = BebopView.getInstance();
		view.startWriting();
		ServerResponseAll.encodeInto(record, view);
		return view.toArray();
	}

	public static encodeInto(record: IServerResponseAll, view: BebopView): number {
		const before = view.length;
		const pos = view.reserveMessageLength();
		const start = view.length;
		if (record.trips !== undefined) {
			view.writeByte(1);
			{
				const length0 = record.trips.length;
				view.writeUint32(length0);
				for (let i0 = 0; i0 < length0; i0++) {
					Trip.encodeInto(record.trips[i0], view);
				}
			}
		}
		view.writeByte(0);
		const end = view.length;
		view.fillMessageLength(pos, end - start);
		const after = view.length;
		return after - before;
	}

	public static decode(buffer: Uint8Array): IServerResponseAll {
		const view = BebopView.getInstance();
		view.startReading(buffer);
		return ServerResponseAll.readFrom(view);
	}

	public static readFrom(view: BebopView): IServerResponseAll {
		let message: IServerResponseAll = {};
		const length = view.readMessageLength();
		const end = view.index + length;
		while (true) {
			switch (view.readByte()) {
				case 0:
					return new ServerResponseAll(message);

				case 1:
					{
						let length0 = view.readUint32();
						message.trips = new Array<ITrip>(length0);
						for (let i0 = 0; i0 < length0; i0++) {
							let x0: ITrip;
							x0 = Trip.readFrom(view);
							message.trips[i0] = x0;
						}
					}
					break;

				default:
					view.index = end;
					return new ServerResponseAll(message);
			}
		}
	}
}